Replace Travis with GitHub Actions (#101)
Currently, we use Travis CI for this repo. This change will move us to GitHub
Actions.

These are some of the things we need to do so we have parity with Travis.

- [x] build
- [x] pass tests
- [x] upload coverage
- [x] release*

goveralls doesn't support github actions.
https://docs.coveralls.io/supported-ci-services

codecov does though! https://docs.codecov.io/docs/supported-ci-providers

This PR switches this repo from coveralls to codecov for coverage reports.

*I need this PR to merge so I can test the release workflow.

This currently tests with Go 1.13 and 1.12.9. In addition, it tests with Linux,
macOS, and Windows.