Add server role to channel metrics to check cardinality
When reading stats like pending count and max in flight it is helpful to filter by
server role != Follower, because followers will report a max in flight but a pending
count of 0.

Example: to determine queue saturation, I would query the following:

```
sum(nss_chan_subs_pending_count) by (channel, queue, durable_name)
  / sum(nss_chan_subs_max_inflight) by (channel, queue, durable_name)
```

But the sum on the right hand side will be too large - Follower nodes
report the same value as the Leader.

Say I have 4 subcribers, each with a max in flight of 10, and a "true"
pending count of 10 messages (divided however), and a cluster size of
5 nodes.

The real "saturation" value should be 25%, (10 pending / (10 max in flight * 4 subscribers))

However with no way to filter out the follower nodes the saturation is reported
as 5%. The numerator is still 10, but the denominator will be `10 * 4 * 5`, the
true max in flight multiplied by subscribers multiplied by servers reporting the
max in flight.