Refactor JetStream account metrics test and update mock data
Enhanced the test to utilize a structured `JSInfo` response instead of a static JSON string, improving the accuracy of the test data. Updated the mock server to handle responses using the new structure and adjusted the expected values accordingly. This change increases the robustness of the JetStream account metrics testing.

Also made logging better

Signed-off-by: Daniel Gedgaudas <daniel.gedgaudas@plejd.com>