chore: Clean up GH actions & go versions
Run zizmor over the GitHub actions and fix issues highlighted.
 * this includes setting dependabot update cooldowns

Pin GitHub Actions using `pinact` (which is compatible with dependeabot).

Bump all go dependencies
 * `go get -u`, `go mod tidy`
 * `go test -v ./...`
 * manually modify the Go version in go.mod, since the release workflow uses
   this version number to decide the Go to use for release builds

Note that the tests passed on the third attempt, the first two flaked out with
being unable to start a NATS server in a Go routine.